Fix motd not displaying with update_motd="NO"
There is no way to have a motd file without a FreeBSD version string appended.
To avoid appending the string, motd rc script should be disabled using
update_motd="NO" and an immutable motd file created.

If the file is created in /var/run it will not survive reboot
because the cleanvar rc script will delete it.
It is also possible to mount /var/run as tmpfs which also won't survive reboot.

Move it from /var/run/motd to /var/db/motd so it will persist after reboot.

PR: bin/250081